FILE: extracts component logic to local ChartsService
